Refactor location for random string generation
