+++ test/test_md5.cpp
@@ -6,6 +6,7 @@
 
 #include <boost/crypt/hash/md5.hpp>
 #include <boost/core/lightweight_test.hpp>
+#include "generate_random_strings.hpp"
 
 #ifdef __clang__
 #  pragma clang diagnostic push
@@ -25,7 +26,6 @@
 #  pragma GCC diagnostic pop
 #endif
 
-
 #include <random>
 #include <iostream>
 #include <string>
@@ -34,27 +34,6 @@
 #include <cstdlib>
 #include <ctime>
 #include <cstring>
-
-void generate_random_cstring(char* str, std::size_t length)
-{
-
-    const char charset[] = "0123456789"
-                           "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
-                           "abcdefghijklmnopqrstuvwxyz";
-
-    const std::size_t charset_size = sizeof(charset) - 1;
-
-    std::mt19937_64 rng(42);
-    std::uniform_int_distribution<int> dist(0, charset_size);
-
-    for (std::size_t i = 0; i < length - 1; ++i)
-    {
-        int index = dist(rng);
-        str[i] = charset[index];
-    }
-
-    str[length - 1] = '\0';
-}
 
 auto get_boost_uuid_result(const char* str, size_t length)
 {
@@ -228,7 +207,7 @@ void test_random_values()
     {
         std::memset(str, '\0', max_str_len);
         const std::size_t current_str_len {str_len(rng)};
-        generate_random_cstring(str, current_str_len);
+        boost::crypt::generate_random_cstring(str, current_str_len);
         const auto uuid_res {get_boost_uuid_result(str, current_str_len)};
         const auto crypt_res {boost::crypt::md5(str, current_str_len)};
 
@@ -265,8 +244,8 @@ void test_random_piecewise_values()
         std::memset(str_2, '\0', max_str_len);
 
         const std::size_t current_str_len {str_len(rng)};
-        generate_random_cstring(str, current_str_len);
-        generate_random_cstring(str_2, current_str_len);
+        boost::crypt::generate_random_cstring(str, current_str_len);
+        boost::crypt::generate_random_cstring(str_2, current_str_len);
 
         boost_hasher.process_bytes(str, current_str_len);
         boost_hasher.process_bytes(str_2, current_str_len);
